+++ src/apps/fuzzers/fuzzerCellProperties.c
@@ -23,6 +23,8 @@
 
 typedef struct {
     H3Index index;
+    H3Index mask;
+    int digit;
 } inputArgs;
 
 int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
@@ -36,6 +38,7 @@ int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
     H3_EXPORT(isValidCell)(args->index);
     H3_EXPORT(isPentagon)(args->index);
     H3_EXPORT(isResClassIII)(args->index);
+    H3_EXPORT(getReservedBits)(args->index);
 
     int faceCount;
     H3Error err = H3_EXPORT(maxFaceCount)(args->index, &faceCount);
@@ -44,6 +47,9 @@ int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
         H3_EXPORT(getIcosahedronFaces)(args->index, out);
         free(out);
     }
+
+    int digitOut;
+    H3_EXPORT(getIndexDigit)(args->index, args->digit, &digitOut);
 
     return 0;
 }